+++ source/Logging/source/RequestResponseMiddleware/IRequestResponseLogging.cs
@@ -30,8 +30,9 @@ namespace Energinet.DataHub.Core.Logging.RequestResponseMiddleware
         /// <param name="metaData">log metaData</param>
         /// <param name="indexTags">index tags</param>
         /// <param name="logName"></param>
+        /// <param name="logFolder"></param>
         /// <returns>work task</returns>
-        Task LogRequestAsync(Stream logStream, Dictionary<string, string> metaData, Dictionary<string, string> indexTags, string logName);
+        Task LogRequestAsync(Stream logStream, Dictionary<string, string> metaData, Dictionary<string, string> indexTags, string logName, string logFolder);
 
         /// <summary>
         /// Logs response
@@ -40,7 +41,8 @@ namespace Energinet.DataHub.Core.Logging.RequestResponseMiddleware
         /// <param name="metaData">log metaData</param>
         /// <param name="indexTags">index tags</param>
         /// <param name="logName"></param>
+        /// <param name="logFolder"></param>
         /// <returns>Work task</returns>
-        Task LogResponseAsync(Stream logStream, Dictionary<string, string> metaData, Dictionary<string, string> indexTags, string logName);
+        Task LogResponseAsync(Stream logStream, Dictionary<string, string> metaData, Dictionary<string, string> indexTags, string logName, string logFolder);
     }
 }