+++ source/Logging/documents/release-notes/release-notes.md
@@ -3,3 +3,12 @@
 ## Version 1.0.0
 
 - Initial release.
+
+## Version 1.0.2
+
+- Changes for log naming and request response stream handling
+
+## Version 1.0.3
+
+- Better naming with subfolder. 
+- Added jwt token parsing for index tags.